fresh branch for e2e tests update
