+++ client/__tests__/e2e/e2e.test.ts
@@ -16,18 +16,55 @@ import {
   getTestId,
   goToPage,
   waitByID,
+  waitByClass,
+  getTestClass,
+  clickOnUntil,
 } from "./puppeteerUtils";
 
 import {
   calcDragCoordinates,
   clip,
   drag,
+  expandCategory,
+  expandGene,
+  assertGenesetDoesNotExist,
+  colorByGeneset,
+  assertGenesetExists,
+  assertColorLegendLabel,
+  editGenesetName,
+  removeGene,
+  colorByGene,
+  addGeneToSetAndExpand,
+  subset,
+  assertGeneDoesNotExist,
+  assertGeneExistsInGeneset,
+  deleteGeneset,
+  createGeneset,
   getAllCategoriesAndCounts,
   getCellSetCount,
   selectCategory,
 } from "./cellxgeneActions";
 
 const data = datasets[DATASET];
+
+const meanExpressionBrushGenesetName = "second_gene_set";
+const genesetToCheckForDescription = "fourth_gene_set";
+const genesetToDeleteName = "geneset_to_delete";
+const editableGenesetName = "geneset_to_edit";
+const editText = "_111";
+const newGenesetName = "geneset_to_edit_111";
+const genesetDescriptionID =
+  "geneset-description-tooltip-fourth_gene_set: fourth description";
+const genesetDescriptionString = "fourth_gene_set: fourth description";
+const geneToAddToSet = "RER1";
+const setToAddGeneTo = "fill_this_geneset";
+const geneToBrushAndColorBy = "SIK1";
+const brushThisGeneGeneset = "brush_this_gene";
+const geneBrushedCellCount = "104";
+const subsetGeneBrushedCellCount = "89";
+const geneToRemove = "SIK1";
+const setToRemoveFrom = "empty_this_geneset";
+const perTestCategoryName = "FAKE";
 
 const defaultBaseUrl = "d";
 const pageUrl = appUrlBase.includes("localhost")
@@ -358,6 +395,231 @@ test("lasso moves after pan", async () => {
   const panCount = await getCellSetCount(2);
 
   expect(panCount).toBe(initialCount);
+});
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any --- FIXME: disabled temporarily on migrate to TS.
+async function setup(config: any) {
+  await goToPage(pageUrl);
+  // page
+  //  .on('console', message =>
+  //    console.log(`${message.type().substr(0, 3).toUpperCase()} ${message.text()}`))
+  if (config.withSubset) {
+    await subset({ x1: 0.1, y1: 0.1, x2: 0.8, y2: 0.8 });
+  }
+}
+
+describe.each([
+  { withSubset: true, tag: "subset" },
+  { withSubset: false, tag: "whole" },
+])("geneSET crud operations and interactions", (config) => {
+  test("brush on geneset mean", async () => {
+    await setup(config);
+    await createGeneset(meanExpressionBrushGenesetName);
+    await addGeneToSetAndExpand(meanExpressionBrushGenesetName, "SIK1");
+
+    const histBrushableAreaId = `histogram-${meanExpressionBrushGenesetName}-plot-brushable-area`;
+
+    const coords = await calcDragCoordinates(histBrushableAreaId, {
+      x1: 0.1,
+      y1: 0.5,
+      x2: 0.9,
+      y2: 0.5,
+    });
+    await drag(histBrushableAreaId, coords.start, coords.end);
+    await clickOn(`cellset-button-1`);
+    const cellCount = await getCellSetCount(1);
+    if (config.withSubset) {
+      expect(cellCount).toBe("113");
+    } else {
+      expect(cellCount).toBe("131");
+    }
+  });
+  test("color by mean expression", async () => {
+    await setup(config);
+    await createGeneset(meanExpressionBrushGenesetName);
+    await addGeneToSetAndExpand(meanExpressionBrushGenesetName, "SIK1");
+
+    await colorByGeneset(meanExpressionBrushGenesetName);
+    await assertColorLegendLabel(meanExpressionBrushGenesetName);
+  });
+  test("diffexp", async () => {
+    if (config.withSubset) return;
+
+    await setup(config);
+
+    // set the two cell sets to b cells vs nk cells
+    await expandCategory(`louvain`);
+    await clickOn(`louvain:category-select`);
+    await clickOn(`categorical-value-select-louvain-B cells`);
+    await clickOn(`cellset-button-1`);
+    await clickOn(`categorical-value-select-louvain-B cells`);
+    await clickOn(`categorical-value-select-louvain-NK cells`);
+    await clickOn(`cellset-button-2`);
+
+    // run diffexp
+    await clickOn(`diffexp-button`);
+    await waitByClass("pop-1-geneset-expand");
+    await expect(page).toClick(getTestClass("pop-1-geneset-expand"));
+
+    await page.waitForFunction(
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any --- FIXME: disabled temporarily on migrate to TS.
+      (selector: any) => !document.querySelector(selector),
+      {},
+      getTestClass("gene-loading-spinner")
+    );
+
+    let genesHTML = await getOneElementInnerHTML(
+      getTestClass("gene-set-genes")
+    );
+
+    expect(genesHTML).toMatchSnapshot();
+
+    await expect(page).toClick(getTestClass("pop-1-geneset-expand"));
+    await expect(page).toClick(getTestClass("pop-2-geneset-expand"));
+
+    await page.waitForFunction(
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any --- FIXME: disabled temporarily on migrate to TS.
+      (selector: any) => !document.querySelector(selector),
+      {},
+      getTestClass("gene-loading-spinner")
+    );
+
+    genesHTML = await getOneElementInnerHTML(getTestClass("gene-set-genes"));
+
+    expect(genesHTML).toMatchSnapshot();
+  });
+  test("create a new geneset and undo/redo", async () => {
+    if (config.withSubset) return;
+
+    await setup(config);
+
+    const genesetName = `test-geneset-foo-123`;
+    await assertGenesetDoesNotExist(genesetName);
+    await createGeneset(genesetName);
+    /* note: as of June 2021, the aria label is in the truncate component which clones the element */
+    await assertGenesetExists(genesetName);
+    await clickOn("undo");
+    await assertGenesetDoesNotExist(genesetName);
+    await clickOn("redo");
+    await assertGenesetExists(genesetName);
+  });
+  test("edit geneset name and undo/redo", async () => {
+    await setup(config);
+    await createGeneset(editableGenesetName);
+    await editGenesetName(editableGenesetName, editText);
+    await assertGenesetExists(newGenesetName);
+    await clickOn("undo");
+    await assertGenesetExists(editableGenesetName);
+    await clickOn("redo");
+    await assertGenesetExists(newGenesetName);
+  });
+  test("delete a geneset and undo/redo", async () => {
+    if (config.withSubset) return;
+
+    await setup(config);
+    await createGeneset(genesetToDeleteName);
+    await deleteGeneset(genesetToDeleteName);
+    await clickOn("undo");
+    await assertGenesetExists(genesetToDeleteName);
+    await clickOn("redo");
+    await assertGenesetDoesNotExist(genesetToDeleteName);
+  });
+  test("geneset description", async () => {
+    if (config.withSubset) return;
+
+    await setup(config);
+    await createGeneset(genesetToCheckForDescription);
+    await clickOnUntil(
+      `${genesetToCheckForDescription}:geneset-expand`,
+      async () => {
+        expect(page).toMatchElement(getTestId(genesetDescriptionID), {
+          text: genesetDescriptionString,
+        });
+      }
+    );
+  });
+});
+
+describe.each([
+  { withSubset: true, tag: "subset" },
+  { withSubset: false, tag: "whole" },
+])("GENE crud operations and interactions", (config) => {
+  test("add a gene to geneset and undo/redo", async () => {
+    await setup(config);
+    await createGeneset(setToAddGeneTo);
+    await addGeneToSetAndExpand(setToAddGeneTo, geneToAddToSet);
+    await assertGeneExistsInGeneset(geneToAddToSet);
+    await clickOn("undo");
+    await assertGeneDoesNotExist(geneToAddToSet);
+    await clickOn("redo");
+    await assertGeneExistsInGeneset(geneToAddToSet);
+  });
+  test("expand gene and brush", async () => {
+    await setup(config);
+    await createGeneset(brushThisGeneGeneset);
+    await addGeneToSetAndExpand(brushThisGeneGeneset, geneToBrushAndColorBy);
+    await expandGene(geneToBrushAndColorBy);
+    const histBrushableAreaId = `histogram-${geneToBrushAndColorBy}-plot-brushable-area`;
+
+    const coords = await calcDragCoordinates(histBrushableAreaId, {
+      x1: 0.25,
+      y1: 0.5,
+      x2: 0.55,
+      y2: 0.5,
+    });
+    await drag(histBrushableAreaId, coords.start, coords.end);
+    const cellCount = await getCellSetCount(1);
+    if (config.withSubset) {
+      expect(cellCount).toBe(subsetGeneBrushedCellCount);
+    } else {
+      expect(cellCount).toBe(geneBrushedCellCount);
+    }
+  });
+  test("color by gene in geneset", async () => {
+    await setup(config);
+    await createGeneset(meanExpressionBrushGenesetName);
+    await addGeneToSetAndExpand(meanExpressionBrushGenesetName, "SIK1");
+
+    await colorByGene("SIK1");
+    await assertColorLegendLabel("SIK1");
+  });
+  test("delete gene from geneset and undo/redo", async () => {
+    // We've already deleted the gene
+    if (config.withSubset) return;
+
+    await setup(config);
+    await createGeneset(setToRemoveFrom);
+    await addGeneToSetAndExpand(setToRemoveFrom, geneToRemove);
+
+    await removeGene(geneToRemove);
+    await assertGeneDoesNotExist(geneToRemove);
+    await clickOn("undo");
+    await assertGeneExistsInGeneset(geneToRemove);
+    await clickOn("redo");
+    await assertGeneDoesNotExist(geneToRemove);
+  });
+});
+
+describe.each([
+  { withSubset: true, tag: "subset", categoricalAnno: true },
+  { withSubset: false, tag: "whole", categoricalAnno: true },
+])("annotations", (config) => {
+  test("stacked bar graph renders", async () => {
+    await setup(config);
+    await expandCategory(perTestCategoryName);
+
+    await clickOn(`colorby-louvain`);
+
+    const labels = await getAllByClass("categorical-row");
+
+    const result = await Promise.all(
+      labels.map((label) =>
+        page.evaluate((element) => element.outerHTML, label)
+      )
+    );
+
+    expect(result).toMatchSnapshot();
+  });
 });
 
 /* eslint-enable no-await-in-loop -- await in loop is needed to emulate sequential user actions */
